LayoutTests/fast/profiler/simple-event-call.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("A simple profile test where an event happens.");
function startTest()
{
    endTest();
}
</script>
</head>

<body onload="startTest()">
This page's JavaScript has an onload event.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  There should be only one onload node.
<div id="output"></div>
</body>
</html>
