LayoutTests/fast/profiler/simple-event-call-expected.txt

CONSOLE MESSAGE: line 59: A simple profile test where an event happens.
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) simple-event-call.html 8
CONSOLE MESSAGE: line 75:   (program) simple-event-call.html 4
CONSOLE MESSAGE: line 75:   onload simple-event-call.html 16
CONSOLE MESSAGE: line 75:    startTest simple-event-call.html 10
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript has an onload event. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. There should be only one onload node.
